backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.9.0
python-dotenv>=1.0.0
openai>=1.6.1
python-multipart>=0.0.9 